[#1437] neofs-cli: Move `util` command to a separate package
